Redirecting where group is showing when logged in
When a group is logging in. If it have a contract, it will be taking to
editContract with those details. If not it will be taking to
newContract.

Removed the backArrow from newContract and editContract if you are a
group.

When you created a contract as a group you are now taken to the home
page.

NewContract now gets a email from the group immediatly and it's not
possible to change it since it's the username for the group.

Updated the interceptor to redirect to newContract if it's a group that
is logged in and the url contains 'contracts'. OBS this needs to be
updated.